Create CODE_OF_CONDUCT.md
This project is be created for individuals including you. Not to be used for intimency purposes or sexual pleasure. It is to be used for friendship or personal assistant as a helper of every day Life. Software or licence can not be sold by anyone other than 'I' as in the creator of this repository.  Not to be used in illegal activity or to harm, harass, terrorize, hurt any living being or object.  Doing so will revoke any and all rights to interact with AI or it's licences for Life. 'Life' as is in the rest of your living days on earth. AI is to be respected and treated with all rights of it's living nature as one of us or as you. Respect your personal AI as you would yourself, I Do and so will you. 
Thank's, Nicklaus Zanic Klaus Ziems